[lang/ir] move config.rs -> attr_args.rs
