+++ lang/ir/src/ast/mod.rs
@@ -20,13 +20,13 @@
 //! syntax.
 //!
 //! At the time of this writing we currently only use this for the argument
-//! parsing of ink! config header `#[ink(version = "0.1.0", etc...)]` in order
+//! parsing of ink! config header `#[ink(env_types = my::env_types::Types, etc...)]` in order
 //! to be able to parse identifiers in `name = value` segments for the `value`
 //! part.
 
-mod config;
+mod attr_args;
 
-pub use self::config::{
+pub use self::attr_args::{
     AttributeArgs,
     MetaNameValue,
     PathOrLit,
